fixed 1822f27 from https://gitee.com/liuyuxiang-bear/graphic_ui/pulls/689 Description: Solve warnings IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I5GPJZ Feature or Bugfix: Bugfix Binary Source:No Signed-off-by: liuyuxiang <liuyuxiang7@huawei.com>

+++ frameworks/common/text.cpp
@@ -518,7 +518,7 @@ uint16_t Text::GetLetterIndexByPosition(const Rect& textRect, const Style& style
     return (lineStart < textLen) ? lineStart : TEXT_ELLIPSIS_END_INV;
 }
 
-void Text::SetAbsoluteSizeSpan(int16_t start, int16_t end, uint8_t size)
+void Text::SetAbsoluteSizeSpan(uint16_t start, uint16_t end, uint8_t size)
 {
 #if ENABLE_VECTOR_FONT
     if (fontId_ == FONT_ID_MAX) {
@@ -546,7 +546,7 @@ void Text::SetAbsoluteSizeSpan(int16_t start, int16_t end, uint8_t size)
     }
 
     if (sizeSpans_ != nullptr && start <= characterSize_) {
-        for (int32_t i = start; i < end && i < characterSize_; i++) {
+        for (uint16_t i = start; i < end && i < characterSize_; i++) {
             sizeSpans_[i].fontId = fontId;
             sizeSpans_[i].size = size;
             sizeSpans_[i].isSizeSpan = true;
